Simplify opinion section
